Creating separate file for metric element tests.
